refactor(app): move doc-entry into docs/
Moves doc-entry.tsx (+css) out of the flat shared/ folder into docs/,
since it's only used by docs/sidebar.tsx and already depended on
docs/generator.ts. Updates import paths accordingly. No behavioral
changes.
